requirements.txt
pandas==1.5.3
paramiko==3.4.0
pysftp==0.2.9
pyodbc==5.0.1
python-dotenv==0.21.1